web: Add support for playing (against) computers
For now the computer just plays the first move it finds. Obviously the
goal is for it to make the best move possible.

I'm not sure if using async for this function is the best.
Multithreading would probably be better, but I want the Promise
interface in JS for now.