cheng: Separate between LegalMove and PseudoMove
Much needed performance boost. Moves generated by us (instead of parsed or manually
crafted) are assumed to be "legal" and avoids a bunch of checks.

There is currently no protection for inserting an illegal `LegalMove` (for example if
the `LegalMove` was from another turn) - that is what the lifetimes were supposed to
achieve in the first place.