cheng: Avoid checking if moves get you out of check
They are legal moves, so if there is a legal move, you are not in checkmate. This
won't have any effect in the perft of the default position, because kings can't get
in check in so little moves. For some reason, the test below is unaffected for depth
4.

$ fen='fen 3r1r2/pQ2n3/1pnk4/2Nppq1p/5bp1/2P3BP/PPN2PP1/1K1R3R w - - 4 25\nperft 4\n'
$ hyperfine "printf '$fen' | ./cheng-base" "printf '$fen' | ./cheng-no-check"

It performs better (~10%) if the perft is done to depth 3.