cheng: Implement Board as an abstraction of BorkedBoard
Dirty commit (not all test pass). Implements an abstraction of BorkedBoard that is
always a valid chess position. This allows us to separate between correctness and
speed, and results in a gigantic speedup (10x-20x) because we don't have to validate
that the move is valid each time it is fed.