cheng: Do not use nightly compiler
It was being annoying cause it was checking all the time if there was newer nightly
versions. Also, if we can avoid it, why not? To be fair, we now have to use vectors
instead of statically allocated arrays, and it takes a noticeable performance hit
when generating precomputed data, especially with the `low_nbits` feature.